feat: migrate Terraform integration tests to Calamari.ExternalTools.Tests
