+++ source/Calamari.ExternalTools.Tests/ExternalTools/Terraform/Azure/example.tfvars
