+++ source/Calamari.ExternalTools.Tests/ExternalTools/Terraform/Azure/versions.tf
