+++ source/Calamari.ExternalTools.Tests/ExternalTools/Terraform/CommonTemplates/SingleVariable.json
