+++ source/Calamari.ExternalTools.Tests/ExternalTools/Terraform/GoogleCloud/example.tfvars
