+++ source/Calamari.ExternalTools.Tests/ExternalTools/Terraform/GoogleCloud/test.txt
