+++ source/Calamari.ExternalTools.Tests/ExternalTools/Terraform/PlanDetailedExitCode/example.tf
