+++ source/Calamari.ExternalTools.Tests/ExternalTools/Terraform/WithOutputSensitiveVariables/example.tf
