+++ source/Calamari.ExternalTools.Tests/ExternalTools/Terraform/WithVariablesSubstitution/example.tf
