+++ source/Calamari.ExternalTools.Tests/ExternalTools/Terraform/WithVariablesSubstitution/terraform.tfvars
