--- source/Calamari.Terraform.Tests/AdditionalParams/example.tf
@@ -1,7 +1,0 @@
-variable "my_var" {
-  description = "the var passed in"
-}
-
-output "my_output" {
-  value = "boo"
-}
