--- source/Calamari.Terraform.Tests/CommonTemplates/InlineJsonWithVariables.json
@@ -1,29 +1,0 @@
-﻿{
-  "variable": {
-    "ami": {
-      "type": "string",
-      "description": "the AMI to use",
-      "default": "1234567890"
-    }
-  },
-  "output": {
-    "test": {
-      "value": "hi there"
-    },
-    "test2": {
-      "value": [
-        "hi there",
-        "hi again"
-      ]
-    },
-    "test3": {
-      "value": "${tomap({ a = \"hi\" })}"
-    },
-    "ami": {
-      "value": "${var.ami}"
-    },
-    "random": {
-      "value": "#{RandomNumber}"
-    }
-  }
-}
