--- source/Calamari.Terraform.Tests/TemplateDirectory/SubFolder/example.tf
@@ -1,3 +1,0 @@
-output "my_output" {
-  value = "boo"
-}
