--- source/Calamari.Terraform.Tests/WithVariables/example.tfvars
@@ -1,1 +1,0 @@
-my_var = "Hello World"
